requirements.txt
torch>=1.7.0
torchvision>=0.8.0
numpy>=1.19.0
Pillow>=8.0.0
scikit-image>=0.18.0
opencv-python>=4.5.0
scipy>=1.6.0
matplotlib>=3.3.0
tensorboard>=2.4.0
tqdm>=4.51.0
kornia>=0.5.0
lpips  # For perceptual loss
PyMatting  # For Laplacian computation 